requirements.txt
click==6.7
Django==2.0.6
django-cors-headers==2.2.0
django-redis==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dlib==19.13.1
face-recognition==1.2.2
face-recognition-models==0.3.0
imageio==2.3.0
jsonfield==2.0.2
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
PyJWT==1.6.4
PyMySQL==0.8.1
pytz==2018.4
requests==2.19.0
urllib3==1.23
uWSGI==2.0.17
eye-game
